requirements.txt
discord.py[voice]
openai==0.27.0
asyncio
yt_dlp
python-dotenv
pydantic
